moved mgt command func to task file
